+++ changelog.md
@@ -1,5 +1,53 @@
 # Changelog
 
+## [5.3.4](https://github.com/vb-consulting/Norm.net/tree/5.3.4) (2023-05-26)
+
+[Full Changelog](https://github.com/vb-consulting/Norm.net/compare/5.3.3...5.3.4)
+
+### New option - `KeepOriginalNames` 
+
+When matching by name, default behavior is to strip all database names of underscores and @ at characters.
+
+This is done to make it easier to map underscore separated names to C# properties (`field_name` to `FieldName` for example).
+
+Set this option to true to skip this behavior and keep original names.
+
+Example:
+
+```csharp
+//
+// in your startup
+//
+NormOptions.Configure(options =>
+{
+    options.KeepOriginalNames = true;
+});
+
+...
+
+class FooBarClass
+{
+    public string? Foo_Bar { get; set; }
+    public string? FooBar { get; set; }
+}
+
+...
+
+result = connection
+    .Read<FooBarClass>("select 'foobar' as foo_bar")
+    .Single();
+
+Assert.Equal("foobar", result.Foo_Bar);
+Assert.Null(result.FooBar);
+```
+
+### New optimization and new performance tests
+
+Some missed performance optimizations were added in this version. See the [changelog](https://github.com/vb-consulting/Norm.net/compare/5.3.3...5.3.4) for more details.
+
+Also, see [performance tests results](https://github.com/vb-consulting/Norm.net/blob/5.3.4/PERFOMANCE-TESTS.md).
+
+
 ## [5.3.3](https://github.com/vb-consulting/Norm.net/tree/5.3.3) (2023-05-16)
 
 [Full Changelog](https://github.com/vb-consulting/Norm.net/compare/5.3.2...5.3.3)